Move action buttons to their own directory
